# Тили-тили/Тили-тили_API_openapi.yaml
openapi: 3.0.3
info:
  title: Тили-тили API
  version: 0.28.0
  description: |
    Контракт бэкенда для PWA «Тили-тили» (tili-tili.ru) — свадебный планировщик.
    Фронтенд (React) уже работает на моках; этот файл — единственный источник правды для бэкенда (Claude Code).

    ## Решения владельца 2026-09-02 (контракт v0.2)
    - Деньги — объект `{ amount: копейки, currency: ISO 4217 }` у КАЖДОЙ суммы.
      Не одна валюта на свадьбу: план §19.7 описывает свадьбу за границей, где
      сделка в евро, а бюджет в рублях. В MVP принимается только RUB.
    - У слота нет собственного статуса. Слот либо пуст, либо несёт сделку;
      состояние живёт в `Deal.state` — шесть значений. `Slot.tileState` —
      производная подпись для мозаики, только для чтения.
    - Гость опознаётся ОДНИМ персональным токеном во всех гостевых путях
      (`/rsvp/{guestToken}`, `/gifts/{guestToken}`, `/join/{guestToken}/…`).
      Общий код свадьбы не годится: по нему нельзя подставить имя в приглашение
      и нельзя дать гостю снять СВОЙ резерв подарка. Анонимность сохраняется —
      система знает гостя, а API пары этот токен не отдаёт никогда (§9).
    - Переноса данных с localStorage-моков не будет: реальных пользователей не
      было, эндпоинт импорта не заводится.

    ## Роли
    - `couple` — молодожёны (владельцы свадьбы): полный доступ.
    - `helper` — помощник (друг/родственник): орг-разделы, без бюджета и сделок.
...
    - `vendor` — подрядчик: свои сделки, лиды, чаты, анкета.
...

...
servers:
...
paths:
...
  /notifications:
    get:
      tags: [notifications]
      summary: Уведомления пользователя
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Notification' }
  /notifications/{id}/read:
    post:
      tags: [notifications]
      summary: Отметить прочитанным
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: OK }


  # ───────────────────────── WISHLIST (пара) ─────────────────────────
...
  /users/me/push-subscriptions:
    post:
      tags: [notifications]
      summary: Зарегистрировать подписку Web Push
      description: |
        VAPID. На iOS работает только с 16.4+ и после установки на экран «Домой» —
        при отказе клиент откатывается на email.

        Пока ключи VAPID не заданы, отвечает 501 `push_not_configured`: принять
        подписку и промолчать значило бы показать «уведомления включены» там,
        где их не будет. Уведомления при этом приходят в приложении —
        `GET /notifications` работает всегда.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, keys]
              properties:
                endpoint: { type: string }
                keys: { type: object }
      responses:
        '201': { description: Создано }
        '501': { $ref: '#/components/responses/NotConfigured' }
    delete:
      tags: [notifications]
      summary: Отписаться от push
      description: 'С параметром endpoint — только подписка этого устройства (её адрес знает только оно); без параметра — все подписки человека.'
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: endpoint, in: query, required: false, schema: { type: string, maxLength: 2048 } }
      responses:
        '204': { description: OK }
...
  /deals/{dealId}/events:
    get:
      tags: [deals]
      summary: Журнал сделки
      description: |
        Что и когда произошло со сделкой: переходы состояния и правки цены.
        Пишется в `deal_events` с самого начала, но прочитать его было негде —
        при споре «мы договаривались на другую сумму» доказательства лежали в
        базе и никому не показывались.

        Видят обе стороны сделки: пара и подрядчик. Автор события назван
        ролью, а не именем — «пара», «подрядчик», «система» (истёкшая мягкая
        бронь — событие без человека).
      security: [{ bearerAuth: [] }]
      parameters:
        - name: dealId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    kind: { type: string, enum: [state, price], description: 'переход состояния или правка цены' }
                    fromState: { $ref: '#/components/schemas/DealState' }
                    toState: { $ref: '#/components/schemas/DealState' }
                    by: { type: string, enum: [couple, vendor, system] }
                    note: { type: string, nullable: true, description: 'для правки цены — старая и новая сумма' }
                    at: { type: string, format: date-time }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /deals/{dealId}:
    patch:
      tags: [bookings]
      summary: Перевести сделку в следующее состояние или исправить сумму
      description: |
        Контракт знал только book/cancel/pay — четыре из шести состояний были
        недостижимы через API. Разрешены только переходы вперёд по цепочке
        candidate → contacted → negotiating → booked → paid_deposit → done
        и cancelled из любого. Недопустимый переход — 409.
        Переход в negotiating ставит negotiatingUntil = now + 72 ч.

        Сумму можно исправить до внесения аванса: подрядчик присылает новую
        смету, и без этой правки пара была вынуждена отменять сделку, а
        отмена освобождает дату в календаре подрядчика. В paid_deposit,
        done и cancelled сумма зафиксирована — 409 `price_locked`.
        Правка пишется в журнал сделки как событие вида `price`.

        Требуется хотя бы одно из полей: `state` или `price`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: dealId, in: path, required: true, schema: { type: string } }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              anyOf: [{ required: [state] }, { required: [price] }]
              properties:
                state: { $ref: '#/components/schemas/DealState' }
                price: { $ref: '#/components/schemas/Money' }
                note: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deal' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Forbidden:
      description: |
        Доступ закрыт роли или стороне. Это нормальный ответ, а не поломка:
        помощник не видит денег, посторонний — чужой сделки (план §6).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Конфликт (уже существует)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotConfigured:
      description: |
        Путь согласован, обработчик есть, но внешняя служба не подключена
        (OAuth-приложения, VAPID-ключи, объектное хранилище). Код ошибки
        называет, чего именно нет: `oauth_not_configured`,
        `push_not_configured`, `storage_not_configured`. Это хвост владельца,
        а не недоделка сервера — 501 с общим `not_implemented` здесь не бывает.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              type: object
              additionalProperties: { type: string }
              description: |
                Только у 422: какие именно поля не прошли проверку.
                Ключ — имя поля (`reason`, `action`, `synonyms.тамада`),
                значение — что с ним не так. Человеку показывается `message`,
                поля подсвечивает форма.
...
    Vendor:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        categoryId: { type: string }
        city: { type: string }
        priceFrom:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          nullable: true
        rating: { type: number, nullable: true, example: 4.9 }
        reviewsCount: { type: integer }
        photoUrl: { type: string, nullable: true }
        verified:
          type: boolean
          description: |
            Документы сверены. Сами документы не публикуются НИКОГДА —
            наружу выходит только этот признак (план §6).
        hasVideo: { type: boolean }
...
    Notification:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        body: { type: string }
        kind: { type: string, enum: [deal, chat, task, guest, system] }
        link:
          type: string
          nullable: true
          example: /deal/01a06eac-4983-7b5e-8d67-36d1a8e05cc9
          description: |
            Место, а не маршрут приложения: `/deal/{dealId}`, `/chats/{chatId}`,
            `/guests`, `/checklist`, `/dayx`, `/vendor-app`. У пары и у
            подрядчика одно и то же место лежит по разным адресам, поэтому
            перевод в маршрут делает клиент. Незнакомое место никуда не ведёт.
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
    Money:
      type: object
      description: |
        Денежная сумма. Решение владельца 2026-09-02: минорные единицы плюс код
        валюты у КАЖДОЙ суммы, а не одна валюта на свадьбу — план §19.7 описывает
        свадьбу за границей, где сделка в евро, а бюджет в рублях.
        В MVP принимается только RUB; enum расширяется вместе с поддержкой курсов.
      required: [amount, currency]
      properties:
        amount:
          type: integer
          minimum: 0
          maximum: 9007199254740991
          description: |
            Сумма в копейках. 85 000 ₽ = 8500000. Дробных значений не бывает.
            Потолок — предел точного целого в JSON: за ним число молча
            округляется, и «потрачено» перестаёт сходиться с суммой сделок.
          example: 8500000
        currency:
          type: string
          description: Код валюты по ISO 4217.
          enum: [RUB]
          example: RUB
    Deal:
      type: object
      description: |
        Договорённость пары с конкретным исполнителем на конкретный слот.
        Носитель состояния: у слота собственного статуса нет (решение владельца
        2026-09-02) — раньше слот, сделка и фронтенд описывали одно понятие тремя
        разными наборами значений.
      properties:
        id: { type: string }
        state: { $ref: '#/components/schemas/DealState' }
        vendor:
          allOf: [{ $ref: '#/components/schemas/Vendor' }]
          nullable: true
          description: null у своего подрядчика не из каталога (§11)
        externalName: { type: string, nullable: true, description: имя своего подрядчика }
        externalPhone: { type: string, nullable: true }
        price: { $ref: '#/components/schemas/Money' }
        paid:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: |
            Сколько уже внесено по этой сделке: платежи `deposit` и `balance`
            минус возвраты, без отменённых. Считается на лету, в базе не
            хранится (§3.1: производное значение расходится с источником).

            Без него «оплачено 30 000 из 50 000» показать нечем, и человек
            держит остаток долга в голове. Уходит вместе с ценой — тому, кто
            видит деньги: помощник и координатор не видят ни того, ни другого.
        paidAt:
          type: string
          format: date-time
          nullable: true
          description: дата последнего платежа. Пусто — платежей не было
        negotiatingUntil:
          type: string
          format: date-time
          nullable: true
          description: 'срок мягкой брони: 72 ч на этапе negotiating (§18.3)'
        bookedAt: { type: string, format: date-time, nullable: true }
        doneAt: { type: string, format: date-time, nullable: true }
        cancelledAt: { type: string, format: date-time, nullable: true }
    DealState:
      type: string
      description: |
        Шесть состояний сделки (План §8.1, решение владельца 2026-09-02).
        Мягкая бронь на 72 часа — не отдельное состояние, а срок жизни `negotiating` (§18.3).
      enum: [candidate, contacted, negotiating, booked, paid_deposit, done, cancelled]
...
security:
  - bearerAuth: []